Replace Nativebase for Forms with TextInputs
This change replaces the Nativebase FormControl and related components
which were used in the NativeLogin and SiteUrl screens. This is done
so we can continue to work towards removing nativebase as it is very outdated
and not used a whole lot.

The change removes the InputTextWithInfo component and replaces it with a
TextInput component which can be dropped into the existing usage without
much change. The floating label is replicated with the Label component, and
error states are managed in the TextInput component with the help of the
Message component to show help/error text