Replace react-native-keyboard-aware-scroll-view with in-house components
This change replaces the react-native-keyboard-aware-scroll-view from the project.
The library was already not really in use - it was only used for the EnrolmentModal but this does
not really need a keyboard avoiding view specifically. The expected behaviour was
achieved here by adding a normal ScrollView and setting some props specific to
keyboard dismissing